refactor : 디렉토리명 통일
