+++ src/main/java/com/sparksInTheStep/webBoard/post/domain/Post.java
@@ -2,7 +2,7 @@ package com.sparksInTheStep.webBoard.post.domain;
 
 import com.sparksInTheStep.webBoard.global.listener.TimeStamp;
 import com.sparksInTheStep.webBoard.member.domain.Member;
-import com.sparksInTheStep.webBoard.post.service.dto.PostCommand;
+import com.sparksInTheStep.webBoard.post.application.dto.PostCommand;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
