+++ src/main/java/com/sparksInTheStep/webBoard/post/presentation/PostController.java
@@ -3,8 +3,8 @@ package com.sparksInTheStep.webBoard.post.presentation;
 import com.sparksInTheStep.webBoard.member.application.dto.MemberInfo;
 import com.sparksInTheStep.webBoard.post.presentation.dto.PostRequest;
 import com.sparksInTheStep.webBoard.post.presentation.dto.PostResponse;
-import com.sparksInTheStep.webBoard.post.service.PostService;
-import com.sparksInTheStep.webBoard.post.service.dto.PostCommand;
+import com.sparksInTheStep.webBoard.post.application.PostService;
+import com.sparksInTheStep.webBoard.post.application.dto.PostCommand;
 import com.sparksInTheStep.webBoard.global.annotation.AuthorizedUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
