+++ src/main/java/com/sparksInTheStep/webBoard/post/presentation/dto/PostResponse.java
@@ -1,7 +1,7 @@
 package com.sparksInTheStep.webBoard.post.presentation.dto;
 
 import com.sparksInTheStep.webBoard.post.domain.PostType;
-import com.sparksInTheStep.webBoard.post.service.dto.PostInfo;
+import com.sparksInTheStep.webBoard.post.application.dto.PostInfo;
 
 import java.time.LocalDateTime;
